refactor : fgetcapabilities added
