+++ wfs_getcapabilities.cpp
@@ -1,19 +1,32 @@
 /* wfs_getcapabilities.cpp
-   Generated by gSOAP 2.8.92 for ws-inspire-wfs2.h
+    Generated for fit2inspire.eu service implementation
 
-gSOAP XML Web services tools
-Copyright (C) 2000-2018, Robert van Engelen, Genivia Inc. All Rights Reserved.
-The soapcpp2 tool and its generated software are released under the GPL.
-This program is released under the GPL with the additional exemption that
-compiling, linking, and/or using OpenSSL is allowed.
---------------------------------------------------------------------------------
-A commercial use license is available from Genivia Inc., contact@genivia.com
---------------------------------------------------------------------------------
+    Copyright(C) 2014-2020, Jean-Loup Delaveau, Fit2Inspire.eu
+
+    This code is released under the following license:
+    1) GPL
 */
 
 
 #include <ctime>
 #include "wfs_getcapabilities.h"
+
+int fgetcapabilities(struct soap *soap)
+{
+    wfs__WFS_x005fCapabilitiesType *wfs__WFS_x005fCapabilities;
+    wfs__WFS_x005fCapabilities = soap_new_wfs__WFS_x005fCapabilitiesType(soap);
+
+    cout << "getCapabilities request received" << endl;
+    __f2i_plu__wfs_x002egetCapabilities(soap, NULL,  *wfs__WFS_x005fCapabilities);
+
+    soap->http_content = "text/xml";
+    soap_response(soap, SOAP_FILE);
+    wfs__WFS_x005fCapabilities->soap_put(soap, "wfs:WFS_Capabilities", "");
+    soap_end_send(soap);
+
+    return SOAP_OK;
+
+} // end function fgetcapabilities
 
 /** Server operation __f2i_plu__wfs_x002egetCapabilities */
 int __f2i_plu__wfs_x002egetCapabilities(struct soap *soap, ows__GetCapabilitiesType *ows__GetCapabilities, wfs__WFS_x005fCapabilitiesType &wfs__WFS_x005fCapabilities)
