+++ wfs_getcapabilities.h
@@ -1,10 +1,8 @@
-/*  ows_schema_instantiate.h
+/*  wfs_getcapabilities.h
     Generated for fit2inspire.eu service implementation
 
     Copyright(C) 2014-2020, Jean-Loup Delaveau, Fit2Inspire.eu
 
-    This code is released under the following license:
-    1) GPL
 */
 
 #ifndef WFS_GETCAPABILITIES_H
@@ -12,7 +10,6 @@
 
 #include "wfs2/src/soapH.h"
 #include "wfs2/ows_schema_instantiate.h"
-#include "wfs_getcapabilities.h"
 #include <mongocxx/client.hpp> // the mongo c++ driver
 #include <mongocxx/uri.hpp>
 #include <mongocxx/instance.hpp>
@@ -27,6 +24,8 @@ using bsoncxx::builder::basic::kvp;
 using bsoncxx::builder::basic::sub_array;
 using bsoncxx::builder::basic::make_document;
 using bsoncxx::oid;
+
+int fgetcapabilities(struct soap *soap);
 
 int http_fget_error(struct soap *soap, const char* exceptionCode, string exceptionText, string locator, int status);
